Move LayoutConfig and RendererLayout to new file
